Refactor secure storage implementations to avoid leaking version data
