+++ consensus/safety-rules/benches/safety_rules.rs
@@ -54,9 +54,10 @@ fn lsr(mut safety_rules: Box<dyn TSafetyRules<Vec<u8>>>, signer: ValidatorSigner
 
 fn in_memory(n: u64) {
     let signer = ValidatorSigner::from_int(0);
-    let storage = InMemoryStorage::new();
-    let storage =
-        PersistentSafetyStorage::initialize(Box::new(storage), signer.private_key().clone());
+    let storage = PersistentSafetyStorage::initialize(
+        InMemoryStorage::new_boxed_in_memory_storage(),
+        signer.private_key().clone(),
+    );
     let safety_rules_manager = SafetyRulesManager::new_local(signer.author(), storage);
     lsr(safety_rules_manager.client(), signer, n);
 }
@@ -64,9 +65,10 @@ fn in_memory(n: u64) {
 fn on_disk(n: u64) {
     let signer = ValidatorSigner::from_int(0);
     let file_path = NamedTempFile::new().unwrap().into_temp_path().to_path_buf();
-    let storage = OnDiskStorage::new(file_path);
-    let storage =
-        PersistentSafetyStorage::initialize(Box::new(storage), signer.private_key().clone());
+    let storage = PersistentSafetyStorage::initialize(
+        OnDiskStorage::new_boxed_on_disk_storage(file_path),
+        signer.private_key().clone(),
+    );
     let safety_rules_manager = SafetyRulesManager::new_local(signer.author(), storage);
     lsr(safety_rules_manager.client(), signer, n);
 }
@@ -74,9 +76,10 @@ fn on_disk(n: u64) {
 fn serializer(n: u64) {
     let signer = ValidatorSigner::from_int(0);
     let file_path = NamedTempFile::new().unwrap().into_temp_path().to_path_buf();
-    let storage = OnDiskStorage::new(file_path);
-    let storage =
-        PersistentSafetyStorage::initialize(Box::new(storage), signer.private_key().clone());
+    let storage = PersistentSafetyStorage::initialize(
+        OnDiskStorage::new_boxed_on_disk_storage(file_path),
+        signer.private_key().clone(),
+    );
     let safety_rules_manager = SafetyRulesManager::new_serializer(signer.author(), storage);
     lsr(safety_rules_manager.client(), signer, n);
 }
@@ -84,9 +87,10 @@ fn serializer(n: u64) {
 fn thread(n: u64) {
     let signer = ValidatorSigner::from_int(0);
     let file_path = NamedTempFile::new().unwrap().into_temp_path().to_path_buf();
-    let storage = OnDiskStorage::new(file_path);
-    let storage =
-        PersistentSafetyStorage::initialize(Box::new(storage), signer.private_key().clone());
+    let storage = PersistentSafetyStorage::initialize(
+        OnDiskStorage::new_boxed_on_disk_storage(file_path),
+        signer.private_key().clone(),
+    );
     let safety_rules_manager = SafetyRulesManager::new_thread(signer.author(), storage);
     lsr(safety_rules_manager.client(), signer, n);
 }
