+++ consensus/safety-rules/src/persistent_safety_storage.rs
@@ -22,7 +22,7 @@ const PREFERRED_ROUND: &str = "preferred_round";
 
 impl PersistentSafetyStorage {
     pub fn in_memory(private_key: Ed25519PrivateKey) -> Self {
-        let storage = Box::new(InMemoryStorage::new());
+        let storage = InMemoryStorage::new_boxed_in_memory_storage();
         Self::initialize(storage, private_key)
     }
 
@@ -109,13 +109,12 @@ impl PersistentSafetyStorage {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use libra_secure_storage::InMemoryStorage;
     use libra_types::crypto_proxies::ValidatorSigner;
 
     #[test]
     fn test() {
         let private_key = ValidatorSigner::from_int(0).private_key().clone();
-        let internal = Box::new(InMemoryStorage::new());
+        let internal = InMemoryStorage::new_boxed_in_memory_storage();
         let mut storage = PersistentSafetyStorage::initialize(internal, private_key);
         assert_eq!(storage.epoch().unwrap(), 1);
         assert_eq!(storage.last_voted_round().unwrap(), 0);
