+++ secure/storage/src/tests/in_memory.rs
@@ -1,10 +1,10 @@
 // Copyright (c) The Libra Core Contributors
 // SPDX-License-Identifier: Apache-2.0
 
-use crate::{in_memory::InMemoryStorage, tests::suite};
+use crate::{tests::suite, InMemoryStorage};
 
 #[test]
 fn in_memory() {
-    let mut storage = Box::new(InMemoryStorage::new());
+    let mut storage = InMemoryStorage::new_boxed_in_memory_storage();
     suite::execute_all_storage_tests(storage.as_mut());
 }
